derphttp_client: add debug envknob to use HTTP client connections
This is a POC for making DERP development easier by allowing clients
to connect with HTTP, thereby avoiding the need for acquiring a valid
TLS cert. This version uses an envknob and hardcodes in the port
number the DERP server uses in "-dev" mode, while acquiring the DERP
server hostname from the control server. To use this:

1. run derper with -dev flag

2. edit the ACL in the admin console to add a custom DERP map with
your dev server's hostname

3. run tailscaled with TS_DEBUG_DERP_USE_HTTP=1

I would prefer to have something akin to "-login-server" for local
control server development, which would instead look like:

1. run derper with -dev flag

2. run tailscaled with -derp-server=localhost

And default to the 3340 port.

Signed-off-by: Val <valerie@tailscale.com>